feat: create useLanguage hook to centralize language logic
